使用Ubuntu上传项目 实现购物车．订单和用户实体类．

+++ shoppingmall-dao/src/main/java/com/lnu/dao/UserRepository.java
@@ -4,7 +4,7 @@ import com.lnu.entity.TbUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
- * @author
+ * @author ccl
  */
-public interface UserRepository extends JpaRepository<TbUser, Integer> {
+public interface UserRepository extends JpaRepository<TbUser, Long> {
 }
